Deleted unused classes and moved another fastfetch-specific class
