+++ GVFS/GVFS.Common/Prefetch/Git/DiffHelper.cs
@@ -1,5 +1,4 @@
-﻿using GVFS.Common;
-using GVFS.Common.Git;
+﻿using GVFS.Common.Git;
 using GVFS.Common.Tracing;
 using System;
 using System.Collections.Concurrent;
